fix: load world size async
